requirements.txt
asgiref==3.5.2
Django==4.1.4
django-filter==22.1
django-markdownfield==0.10.0
djangorestframework==3.14.0
Markdown==3.4.1
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3